Make network_vlan id parameter the namevar instead of name
Without this patch the name parameter is the namevar.  Based on feedback
from vendors, the VLAN id should be the namevar in order to reduce the
number of API calls necessary to inspect and manage a single VLAN.  If
the ID is not the namevar then all resources must be collected and the
name scanned to extract the resource information from the system.
Making the id the namevar also follows the advice of Puppet Types and
Providers where the namevar should be knowable in advance by the end
users.

This patch addressees the problem by changing the namevar from name to
id and renaming the name attribute to vlan_name.  Because ID is the
namevar, name is an invalid attribute for the vlan name because it is
automatically mapped to id by the RAL.