Remove default values for network_interface type
Without this patch there are default values for the network_interface
type.  This is a problem because these default values interfere with
puppet resource's ability to manage only those properties specified by
the end user.

This patch also removes the ensure parameter because physical interfaces
cannot be created or destroyed by software, they always exist or do not
exist.  The enable property replaces ensure to manage the basic state of
shutdown or not shutdown.