(maint) Fix all rubocop warnings and errors
Without this patch there are hundreds of warnings flagged by running
`bundle exec rubocop -FD`  This is a problem because the standard types
need to follow a consistent style.  This patch addresses the problem by
modifying all of the code to make sure rubocop passes without any
warnings or errors.