(maint) Add tasks for puppetmodule-netdev_stdlib gem
Without this patch the netdev_stdlib module cannot be released and
installed as a Ruby gem.  This is a problem because it is difficult to
write a provider module that loads the type definitions in the test
framework.  This problem could be addressed by configuring each provider
module to initialize Puppet's module loading directories using the
spec_helper, but each module would need to do this in a consistent way.
This consistency could be provided by the puppetlabs_spec_helper, but
this too would need to come up with some way to generically load
dependant modules into the load path.

This patch addresses the problem by releasing the dependency as a Gem.
This, in turn, allows a provider module to simply express the dependency
on the netdev_stdlib module as a dependency on the
puppetmodule-netdev_stdlib gem in the Gemfile of the provider module.
Bundler will then easily manage the dependency and can enable satisfying
the dependency from a local git clone.