(maint) Add dependencies back to Gemfile
Without this patch travis is failing because it installs dependencies
using the --without development flag and the rspec dependencies are
flagged as development dependencies in the gemspec.

This patch addresses the problem by adding the testing dependencies to
the Gemfile in the :test and :development groups.

See https://travis-ci.org/puppetlabs/netdev_stdlib/jobs/34857232 for the
failure.