(NETDEV-18) Make network_snmp not ensurable
While implementing the network_snmp type I've run into new information
that there will really only ever be a single resource of this type. As a
result, there isn't really a need to make the type ensurable because the
global SNMP settings for the device are only enabled or disabled, not
added or removed. Because the resource is always "present" the type
should have the ensure attribute removed.