(NETDEV-22) Change snmp_user to composite namevar on name, version
Without this patch the snmp_user type identifies resources by name only.
This is a problem because the same user name can exist multiple times on
the target device provided each user is for a different SNMP version.
This patch addresses the problem by changing the resource model to use
version as a namevar component.