Add hostname property to radius_server type
Without this patch the radius_server type is difficult to manage because
there are three properties, hostname, accounting port, and
authentication port, that determine uniqueness of the resource.  A
composite namevar is not a good solution to this difficulty because it
obscures the accounting port and authentication port by forcing them to
be parameters rather than properties.

This patch addresses the problem by adding an additional property,
hostname, to manage the IP address or hostname value of the radius
server.  The name parameter then becomes only a unique identifier and
not something that's actually managed on the target device.