(NETDEV-23) Add network_dns type
Without this patch there is no type to manage the DNS configuration of a
target device.  This patch addresses the problem by adding the
network_dns type with a namevar of `name` (only used to identify the
resources in the Puppet catalog), a string property for the domain, and
array properties of `search` and `servers`.