(NETDEV-23) Improve network_dns change messages
Without this patch the network_dns change messages look like strings
when comparing the servers and search properties, which are modeled as
Array's.  This patch addresses the problem by updating the formatting
methods for the search and servers properties to display Array's using
Puppet's format_value_for_display method.

This patch also updates the spec tests to use shared behaviors.